Refactor: Move desktop shortcut initialization to files store
