+++ src/components/layout/Desktop.tsx
@@ -503,121 +503,6 @@ export function Desktop({
         )
       : sortedApps;
 
-  // Create default shortcuts based on theme
-  useEffect(() => {
-    // Ensure apps are loaded and Desktop folder exists
-    if (!apps || apps.length === 0) return;
-    
-    const desktopFolder = fileStore.getItem("/Desktop");
-    if (!desktopFolder || !desktopFolder.isDirectory) {
-      // Desktop folder doesn't exist yet, wait for it to be initialized
-      return;
-    }
-
-    // Get current desktop shortcuts (active) and trashed items
-    const desktopItems = fileStore.getItemsInPath("/Desktop");
-    const trashedItems = fileStore.getTrashItems();
-    
-    // Define the default order for desktop shortcuts
-    const defaultOrder: AppId[] = [
-      "ipod",
-      "chats",
-      "applet-viewer",
-      "internet-explorer",
-      "textedit",
-      "photo-booth",
-      "videos",
-      "paint",
-      "soundboard",
-      "minesweeper",
-      "synth",
-      "terminal",
-      "pc",
-    ];
-
-    // Determine which apps should have shortcuts based on theme
-    let appsToShortcut: typeof apps;
-    if (currentTheme === "macosx") {
-      // macOS X: only iPod and Applet Store (applet-viewer) as default shortcuts
-      appsToShortcut = apps.filter(
-        (app) => app.id === "ipod" || app.id === "applet-viewer"
-      );
-    } else {
-      // Other themes: all apps except Finder and Control Panels
-      appsToShortcut = apps.filter(
-        (app) => app.id !== "finder" && app.id !== "control-panels"
-      );
-    }
-
-    // Sort apps according to default order, then alphabetically for any not in the list
-    const sortedAppsToShortcut = [...appsToShortcut].sort((a, b) => {
-      const aIndex = defaultOrder.indexOf(a.id as AppId);
-      const bIndex = defaultOrder.indexOf(b.id as AppId);
-      
-      // If both are in the order list, sort by their position
-      if (aIndex !== -1 && bIndex !== -1) {
-        return aIndex - bIndex;
-      }
-      // If only one is in the list, prioritize it
-      if (aIndex !== -1) return -1;
-      if (bIndex !== -1) return 1;
-      // If neither is in the list, sort alphabetically
-      return a.name.localeCompare(b.name);
-    });
-
-    // Create shortcuts for apps that don't have them yet, in the specified order.
-    // We consider both active and previously trashed shortcuts so that if the
-    // user deletes a default icon, it is not auto-added back on theme changes.
-    sortedAppsToShortcut.forEach((app) => {
-      const appId = app.id as AppId;
-
-      const hasActiveShortcut = desktopItems.some(
-        (item) => item.aliasType === "app" && item.aliasTarget === appId
-      );
-
-      const hasTrashedShortcut = trashedItems.some(
-        (item) =>
-          item.aliasType === "app" &&
-          item.aliasTarget === appId &&
-          item.originalPath?.startsWith("/Desktop/")
-      );
-
-      if (hasActiveShortcut || hasTrashedShortcut) {
-        return;
-      }
-
-      // Use app path if available
-      const appPath = `/Applications/${app.name}`;
-      fileStore.createAlias(appPath, app.name, "app", appId);
-
-      // Mark default shortcuts as theme-conditional by hiding them on macOS X
-      // unless they are iPod or Applet Store, which should always be visible.
-      const latestDesktopItems = fileStore.getItemsInPath("/Desktop");
-      const createdShortcut = latestDesktopItems.find(
-        (item) =>
-          item.aliasType === "app" &&
-          item.aliasTarget === appId &&
-          item.status === "active"
-      );
-
-      if (!createdShortcut) {
-        return;
-      }
-
-      // Only apply theme-conditional hiding for non-macOS themes and only for
-      // apps that are NOT iPod or Applet Store.
-      if (
-        currentTheme !== "macosx" &&
-        appId !== "ipod" &&
-        appId !== "applet-viewer"
-      ) {
-        fileStore.updateItemMetadata(createdShortcut.path, {
-          hiddenOnThemes: ["macosx"],
-        });
-      }
-    });
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [apps, currentTheme]);
 
   const getContextMenuItems = (): MenuItem[] => {
     if (contextMenuShortcutPath) {
